Changes to SignUp Control.
Moved sign up dto from SignUp.razor.cs to models folder. Started to manually build the user to be created before submitting to the create routine
